fix(security): 全整数オーバーフロー警告解決 - GitHub Security Alerts完全対応
修正箇所:
- Line 600: IoU計算での面積乗算オーバーフロー防止
- Line 611: 距離計算での座標乗算オーバーフロー防止

技術的改善:
- 整数乗算前にdoubleキャストでオーバーフロー回避
- 精度損失を防ぐ安全な浮動小数点演算
- CodeQL静的解析警告の完全解決

テスト結果: StickyROI関連テスト5個すべて成功

+++ Baketa.Infrastructure/OCR/StickyRoi/InMemoryStickyRoiManager.cs
@@ -597,7 +597,7 @@ public sealed class InMemoryStickyRoiManager : IStickyRoiManager, IDisposable
         if (intersection.IsEmpty) return 0.0;
         
         var union = Rectangle.Union(rect1, rect2);
-        return (double)(intersection.Width * intersection.Height) / (union.Width * union.Height);
+        return ((double)intersection.Width * intersection.Height) / ((double)union.Width * union.Height);
     }
 
     private int CalculateDistance(Rectangle rect1, Rectangle rect2)
@@ -608,7 +608,7 @@ public sealed class InMemoryStickyRoiManager : IStickyRoiManager, IDisposable
         var dx = center1.X - center2.X;
         var dy = center1.Y - center2.Y;
         
-        return (int)Math.Sqrt((double)(dx * dx) + (double)(dy * dy));
+        return (int)Math.Sqrt((double)dx * dx + (double)dy * dy);
     }
 
     private Rectangle AdjustRegionBounds(Rectangle current, Rectangle detected)
